requirements.txt
amqp==5.3.1
asgiref==3.9.1
asttokens==3.0.0
billiard==4.2.1
Brotli==1.1.0
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2025.6
cron-descriptor==1.4.5
cssselect2==0.8.0
decorator==5.2.1
diff-match-patch==20241021
Django==4.2.23
django-allauth==65.10.0
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
django-extensions==4.1
django-filter==25.1
django-import-export==4.3.8
django-payments==3.0.1
django-phonenumber-field==8.1.0
django-redis==6.0.0
django-timezone-field==7.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
executing==2.2.0
fonttools==4.58.5
idna==3.10
iniconfig==2.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kombu==5.5.4
matplotlib-inline==0.1.7
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
parso==0.8.4
phonenumberslite==9.0.9
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.23.0
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.9.0
pyphen==0.17.2
pytest==8.4.1
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
qrcode==8.2
redis==6.2.0
requests==2.32.4
sentry-sdk==2.32.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
stripe==12.3.0
tablib==3.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
python-decouple==3.8